The Elder Scrolls III: Morrowind — Update script to 2.11 syntax

+++ play.it-2/games/play-the-elder-scrolls-3-morrowind.sh
@@ -1,4 +1,4 @@
-#!/bin/sh -e
+#!/bin/sh
 set -o errexit
 
 ###
@@ -30,11 +30,11 @@ set -o errexit
 
 ###
 # The Elder Scrolls III: Morrowind
-# build native Linux packages from the original installers
+# build native packages from the original installers
 # send your bug reports to vv221@dotslashplay.it
 ###
 
-script_version=20190402.1
+script_version=20190402.2
 
 # Set game-specific variables
 
@@ -56,22 +56,22 @@ ARCHIVE_GOG_FR_VERSION='1.6.1820-gog2.0.0.7'
 ARCHIVE_GOG_FR_SIZE='2300000'
 
 ARCHIVE_DOC_DATA_PATH='app'
-ARCHIVE_DOC_DATA_FILES='./*.pdf'
+ARCHIVE_DOC_DATA_FILES='*.pdf'
 
 ARCHIVE_DOC_L10N_PATH='app'
-ARCHIVE_DOC_L10N_FILES='./*.txt'
+ARCHIVE_DOC_L10N_FILES='*.txt'
 
 ARCHIVE_GAME_BIN_PATH='app'
-ARCHIVE_GAME_BIN_FILES='./*.exe ./binkw32.dll ./morrowind.ini ./tes?construction?set.cnt ./tes?construction?set.hlp'
+ARCHIVE_GAME_BIN_FILES='*.exe binkw32.dll morrowind.ini tes?construction?set.cnt tes?construction?set.hlp'
 
 ARCHIVE_GAME_L10N_PATH='app'
-ARCHIVE_GAME_L10N_FILES='./data?files/bookart/*_377_253.tga ./data?files/bookart/empire?small*.bmp ./data?files/*.bsa ./data?files/*.esm ./data?files/sound/vo ./data?files/splash ./data?files/textures/menu_credits* ./data?files/textures/menu_*game* ./data?files/textures/menu_options* ./data?files/textures/menu_return* ./data?files/textures/tx_menubook_cancel* ./data?files/textures/tx_menubook_close* ./data?files/textures/tx_menubook_journal* ./data?files/textures/tx_menubook_next* ./data?files/textures/tx_menubook_prev* ./data?files/textures/tx_menubook_take* ./data?files/textures/tx_menubook_topics* ./data?files/video/bethesda?logo.bik ./data?files/video/bm_bearhunt?.bik ./data?files/video/bm_ceremony?.bik ./data?files/video/bm_endgame.bik ./data?files/video/bm_frostgiant?.bik ./data?files/video/mw_cavern.bik ./data?files/video/mw_credits.bik ./data?files/video/mw_end.bik ./data?files/video/mw_intro.bik ./data?files/video/mw_logo.bik ./data?files/meshes/r/*atronach_frost.*'
+ARCHIVE_GAME_L10N_FILES='data?files/bookart/*_377_253.tga data?files/bookart/empire?small*.bmp data?files/*.bsa data?files/*.esm data?files/sound/vo data?files/splash data?files/textures/menu_credits* data?files/textures/menu_*game* data?files/textures/menu_options* data?files/textures/menu_return* data?files/textures/tx_menubook_cancel* data?files/textures/tx_menubook_close* data?files/textures/tx_menubook_journal* data?files/textures/tx_menubook_next* data?files/textures/tx_menubook_prev* data?files/textures/tx_menubook_take* data?files/textures/tx_menubook_topics* data?files/video/bethesda?logo.bik data?files/video/bm_bearhunt?.bik data?files/video/bm_ceremony?.bik data?files/video/bm_endgame.bik data?files/video/bm_frostgiant?.bik data?files/video/mw_cavern.bik data?files/video/mw_credits.bik data?files/video/mw_end.bik data?files/video/mw_intro.bik data?files/video/mw_logo.bik data?files/meshes/r/*atronach_frost.*'
 
 ARCHIVE_GAME_DATA_PATH='app'
-ARCHIVE_GAME_DATA_FILES='./data?files/bookart/barbarian_*.tga ./data?files/bookart/boethiah_256.tga ./data?files/bookart/divinemetaphysics_text?.tga ./data?files/bookart/divinemetaphysics.tga ./data?files/bookart/efoulkefirmament_*.tga ./data?files/bookart/eggoftime_illust?.tga ./data?files/bookart/*.htm ./data?files/bookart/magicstonemap4.dds ./data?files/bookart/moragtong.tga ./data?files/bookart/secret_of_dwemer?.tga ./data?files/bookart/*.ttf ./data?files/bookart/tx_icon_waterbreath.bmp ./data?files/*.esp ./data?files/fonts ./data?files/icons ./data?files/meshes ./data?files/music ./data?files/sound/cr ./data?files/sound/fx ./data?files/textures ./data?files/*.txt ./data?files/video/bm_were*.bik ./data?files/video/mw_menu.bik ./knife.ico'
+ARCHIVE_GAME_DATA_FILES='data?files/bookart/barbarian_*.tga data?files/bookart/boethiah_256.tga data?files/bookart/divinemetaphysics_text?.tga data?files/bookart/divinemetaphysics.tga data?files/bookart/efoulkefirmament_*.tga data?files/bookart/eggoftime_illust?.tga data?files/bookart/*.htm data?files/bookart/magicstonemap4.dds data?files/bookart/moragtong.tga data?files/bookart/secret_of_dwemer?.tga data?files/bookart/*.ttf data?files/bookart/tx_icon_waterbreath.bmp data?files/*.esp data?files/fonts data?files/icons data?files/meshes data?files/music data?files/sound/cr data?files/sound/fx data?files/textures data?files/*.txt data?files/video/bm_were*.bik data?files/video/mw_menu.bik knife.ico'
 
 ARCHIVE_GAME_DATAFILES_DATA_PATH='app/_officialplugins/_unpacked_files'
-ARCHIVE_GAME_DATAFILES_DATA_FILES='./*'
+ARCHIVE_GAME_DATAFILES_DATA_FILES='*'
 
 CONFIG_FILES='./*.ini'
 DATA_DIRS='./saves'
@@ -80,7 +80,6 @@ DATA_FILES='./ProgramFlow.txt ./Warnings.txt ./Journal.htm'
 APP_MAIN_TYPE='wine'
 APP_MAIN_EXE='morrowind launcher.exe'
 APP_MAIN_ICON='morrowind.exe'
-APP_MAIN_ICON_RES='16 32'
 
 PACKAGES_LIST='PKG_BIN PKG_L10N PKG_DATA'
 
@@ -107,40 +106,45 @@ PKG_BIN_DESCRIPTION_GOG_FR='French version'
 
 # Load common functions
 
-target_version='2.4'
+target_version='2.11'
 
 if [ -z "$PLAYIT_LIB2" ]; then
-	[ -n "$XDG_DATA_HOME" ] || XDG_DATA_HOME="$HOME/.local/share"
-	if [ -e "$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh" ]; then
-		PLAYIT_LIB2="$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh"
-	elif [ -e './libplayit2.sh' ]; then
-		PLAYIT_LIB2='./libplayit2.sh'
-	else
-		printf '\n\033[1;31mError:\033[0m\n'
-		printf 'libplayit2.sh not found.\n'
-		exit 1
-	fi
+	: "${XDG_DATA_HOME:="$HOME/.local/share"}"
+	for path in\
+		"$PWD"\
+		"$XDG_DATA_HOME/play.it"\
+		'/usr/local/share/games/play.it'\
+		'/usr/local/share/play.it'\
+		'/usr/share/games/play.it'\
+		'/usr/share/play.it'
+	do
+		if [ -e "$path/libplayit2.sh" ]; then
+			PLAYIT_LIB2="$path/libplayit2.sh"
+			break
+		fi
+	done
 fi
-#shellcheck source=play.it-2/lib/libplayit2.sh
+if [ -z "$PLAYIT_LIB2" ]; then
+	printf '\n\033[1;31mError:\033[0m\n'
+	printf 'libplayit2.sh not found.\n'
+	exit 1
+fi
+# shellcheck source=play.it-2/lib/libplayit2.sh
 . "$PLAYIT_LIB2"
 
 # Extract game data
 
 extract_data_from "$SOURCE_ARCHIVE"
-
-for PKG in $PACKAGES_LIST; do
-	organize_data "DOC_${PKG#PKG_}"  "$PATH_DOC"
-	organize_data "GAME_${PKG#PKG_}" "$PATH_GAME"
-done
-
+prepare_package_layout
 PKG='PKG_DATA'
 organize_data 'GAME_DATAFILES_DATA' "$PATH_GAME/data files"
+rm --recursive "$PLAYIT_WORKDIR/gamedata"
+
+# Extract icons
 
 PKG='PKG_BIN'
-extract_and_sort_icons_from 'APP_MAIN'
-move_icons_to 'PKG_DATA'
-
-rm --recursive "$PLAYIT_WORKDIR/gamedata"
+icons_get_from_package 'APP_MAIN'
+icons_move_to 'PKG_DATA'
 
 # Fix .bsa/.esm dates on French version
 
@@ -159,7 +163,7 @@ fi
 # Write launchers
 
 PKG='PKG_BIN'
-write_launcher 'APP_MAIN'
+launchers_write 'APP_MAIN'
 
 # Build package
